chk prev image first before delete

+++ app/controllers/api/users_controller.rb
@@ -461,22 +461,20 @@ class Api::UsersController < ApplicationController
 
         if params[:avatar_url].present?
 
-          if Rails.env.development?
-            bucket_name = AWS_Bucket::Avatar_D
-          elsif Rails.env.staging?
-            bucket_name = AWS_Bucket::Avatar_S
-          else
-            bucket_name = AWS_Bucket::Avatar_P
+          if !current_user.avatar_url.url.nil?
+            if Rails.env.development?
+              bucket_name = AWS_Bucket::Avatar_D
+            elsif Rails.env.staging?
+              bucket_name = AWS_Bucket::Avatar_S
+            else
+              bucket_name = AWS_Bucket::Avatar_P
+            end
+            Post.delete_S3_file(bucket_name, user.avatar_url.current_path,Post::IMAGE)
           end
-          Post.delete_S3_file(bucket_name, user.avatar_url.current_path,Post::IMAGE)
+
           user.remove_avatar_url!
           user.avatar_url = params[:avatar_url]
-          
-          # if params[:avatar_url].to_s == "null"
-          #
-          # else
-          #
-          # end
+
         end
 
         user.save!
